nhsocket-master/requirements.txt
anyio==4.4.0
astroid==3.3.2
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
dill==0.3.7
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
isort==5.12.0
mccabe==0.7.0
platformdirs==4.2.2
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.0.7
websockets==13.0.1
